==> inst/3_model_tables.Rmd <==
---
title: "Tables model output WMT XER"
author: "C. Emi Fergus"
date: "2023-06-26"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Packages and data
```{r, echo=FALSE, message=FALSE}
library(dplyr)
library(tidyr)
library(kableExtra)
library(knitr)
library(stringr)
library(readr)

library(devtools)
devtools::load_all()
library(SEMWesternStrmbug)


###################
# SEM Standardized OUTPUT
# ECO9 WMT
## OE
WMTw_m15_OE_CI <- read_csv("Routput/WMTw_m15_OE_CI.csv")
#WMTw_m15_OE_CI<-read_csv("Routput/WMTw_m15_OE_CI.csv")

## MMI
WMTw_m15_MMI_CI<-read_csv("Routput/WMTw_m15_MMI_CI.csv")

## EPT
WMTw_m15_EPT_CI<-read_csv("Routput/WMTw_m15_EPT_CI.csv")

########################
# ECO9 XER
## OE
XERw_m15_OE_CI<-read_csv("Routput/XERw_m15_OE_CI.csv")

## MMI
XERw_m15_MMI_CI<-read_csv("Routput/XERw_m15_MMI_CI.csv")

## EPT
XERw_m15_EPT_CI<-read_csv("Routput/XERw_m15_EPT_CI.csv")

```

# Processing estimated model effects
```{r, echo=FALSE, message=FALSE}

devtools::load_all()

# Call function sem_eff_tab to process the raw SEM output
# Selects Total, Direct, and Indirect Effects on OE
# And relabels predictor variables to more useful names
# O/E
oe_proc<- sem_eff_tab(WMTw_m15_OE_CI)
names(oe_proc)

# MMI
mmi_proc<-sem_eff_tab(WMTw_m15_MMI_CI)

# EPT
ept_proc<-sem_eff_tab(WMTw_m15_EPT_CI)

######
## XER
# O/E
oe_x_proc<- sem_eff_tab(XERw_m15_OE_CI)
names(oe_proc)

# MMI
mmi_x_proc<-sem_eff_tab(XERw_m15_MMI_CI)

# EPT
ept_x_proc<-sem_eff_tab(XERw_m15_EPT_CI)

```

## Table Full WMT OE
```{r, echo=FALSE}
kable(oe_proc, caption = "WMT OE: Standardized Direct, Indirect, and Total Effects on benthic O/E")%>%
  kable_styling()
```

## Table MMI
```{r, echo=FALSE}
kable(mmi_proc, caption = "WMT MMI: Standardized Direct, Indirect, and Total Effects on benthic MMI")%>%
  kable_styling()
```

## Table EPT
```{r, echo=FALSE}
kable(ept_proc, caption = "WMT EPT: Standardized Direct, Indirect, and Total Effects on benthic EPT")%>%
  kable_styling()
```


## Table Full XER OE
```{r, echo=FALSE}
kable(oe_x_proc, caption = "XER OE: Standardized Direct, Indirect, and Total Effects on benthic O/E")%>%
  kable_styling()
```

## Table MMI
```{r, echo=FALSE}
kable(mmi_x_proc, caption = "XER MMI: Standardized Direct, Indirect, and Total Effects on benthic MMI")%>%
  kable_styling()
```

## Table EPT
```{r, echo=FALSE}
kable(ept_x_proc, caption = "Xer EPT: Standardized Direct, Indirect, and Total Effects on benthic EPT")%>%
  kable_styling()
```


## Model Fit Tables
```{r, echo=FALSE}
# LOAD SEM MODEL FIT TABLES
WMTw_m15_OE_fit<-read_csv("Routput/WMTw_m15_OE_fit.csv")

WMTw_m15_MMI_fit<-read_csv("Routput/WMTw_m15_MMI_fit.csv")

WMTw_m15_EPT_fit<-read_csv("Routput/WMTw_m15_EPT_fit.csv")

fit<-bind_rows(WMTw_m15_OE_fit,WMTw_m15_MMI_fit, WMTw_m15_EPT_fit)

kable(fit, caption = "Model fit parameters for WMT")%>%
  kable_styling()

```

## XER Model Fit Tables
```{r, echo=FALSE}
# LOAD SEM MODEL FIT TABLES
XERw_m15_OE_fit<-read_csv("Routput/XERw_m15_OE_fit.csv")

XERw_m15_MMI_fit<-read_csv("Routput/XERw_m15_MMI_fit.csv")

XERw_m15_EPT_fit<-read_csv("Routput/XERw_m15_EPT_fit.csv")

fit_x<-bind_rows(XERw_m15_OE_fit,XERw_m15_MMI_fit, XERw_m15_EPT_fit)

kable(fit_x, caption = "Model fit parameters for XER")%>%
  kable_styling()

```

# R2 Tables
```{r, echo=FALSE}
# Read in R2 tables for all endogenous variables
# WEST O/E
WMTw_m15_OE_R2<-read_csv("Routput/WMTw_m15_OE_R2.csv")

WMTw_m15_MMI_R2<-read_csv("Routput/WMTw_m15_MMI_R2.csv")

WMTw_m15_EPT_R2<-read_csv("Routput/WMTw_m15_EPT_R2.csv")

# Apply R2 table rename function
R2_wmt_oe_mod<-r2_tab_v2(WMTw_m15_OE_R2)
R2_wmt_mmi_mod<-r2_tab_v2(WMTw_m15_MMI_R2)
R2_wmt_ept_mod<-r2_tab_v2(WMTw_m15_EPT_R2)

# Modify tables and then bring together
R2_wmt_oe_mod2<-R2_wmt_oe_mod%>%
  mutate(Model = "OE")%>%
  select(Model,Response,R2)%>%
  #rename(R2 = est, Response=lhs)%>%
  mutate(across(where(is.numeric), round,2))

R2_wmt_mmi_mod2<-R2_wmt_mmi_mod%>%
  mutate(Model = "MMI")%>%
  select(Model,Response,R2)%>%
  #rename(R2 = est, Response=lhs)%>%
  mutate(across(where(is.numeric), round,2))

R2_wmt_ept_mod2<-R2_wmt_ept_mod%>%
  mutate(Model = "EPT")%>%
  select(Model,Response,R2)%>%
  #rename(R2 = est, Response=lhs)%>%
  mutate(across(where(is.numeric), round,2))

r2_all<-rbind(R2_wmt_oe_mod2,R2_wmt_mmi_mod2, R2_wmt_ept_mod2)

```

## Table all R2 WMT
```{r, echo=FALSE}
kable(r2_all, caption = "R2 estimates across models WMT")%>%
  kable_styling()
```

## R2 XER 
```{r, echo=FALSE}
# Read in R2 tables for all endogenous variables
# XER O/E
XERw_m15_OE_R2<-read_csv("Routput/XERw_m15_OE_R2.csv")

XERw_m15_MMI_R2<-read_csv("Routput/XERw_m15_MMI_R2.csv")

XERw_m15_EPT_R2<-read_csv("Routput/XERw_m15_EPT_R2.csv")

# Apply R2 table rename function
R2_xer_oe_mod<-r2_tab_v2(XERw_m15_OE_R2)
R2_xer_mmi_mod<-r2_tab_v2(XERw_m15_MMI_R2)
R2_xer_ept_mod<-r2_tab_v2(XERw_m15_EPT_R2)

# Modify tables and then bring together
R2_xer_oe_mod2<-R2_xer_oe_mod%>%
  mutate(Model = "OE")%>%
  select(Model,Response,R2)%>%
  #rename(R2 = est, Response=lhs)%>%
  mutate(across(where(is.numeric), round,2))

R2_xer_mmi_mod2<-R2_xer_mmi_mod%>%
  mutate(Model = "MMI")%>%
  select(Model,Response,R2)%>%
  #rename(R2 = est, Response=lhs)%>%
  mutate(across(where(is.numeric), round,2))

R2_xer_ept_mod2<-R2_xer_ept_mod%>%
  mutate(Model = "EPT")%>%
  select(Model,Response,R2)%>%
  #rename(R2 = est, Response=lhs)%>%
  mutate(across(where(is.numeric), round,2))

r2_all_x<-rbind(R2_xer_oe_mod2,R2_xer_mmi_mod2, R2_xer_ept_mod2)

```

## Table XER R2
```{r, echo=FALSE}
kable(r2_all_x, caption = "R2 estimates across models XER")%>%
  kable_styling()
```
